+++ Yvand.EntraCP.Tests/FilterAccountEnabledTests.cs
@@ -1,0 +1,37 @@
+﻿using NUnit.Framework;
+using System;
+
+namespace Yvand.EntraClaimsProvider.Tests
+{
+    [TestFixture]
+    [Parallelizable(ParallelScope.Children)]
+    public class FilterAccountEnabledTests : ClaimsProviderTestsBase
+    {
+        public override void InitializeSettings()
+        {
+            base.InitializeSettings();
+            Settings.FilterAccountsEnabledOnly = true;
+            base.ApplySettings();
+        }
+
+        [Test]
+        public override void CheckSettingsTest()
+        {
+            base.CheckSettingsTest();
+        }
+
+        [Test, TestCaseSource(typeof(TestEntitySourceManager), nameof(TestEntitySourceManager.GetSomeUsers), new object[] { TestEntitySourceManager.MaxNumberOfUsersToTest })]
+        public void TestUsers(TestUser user)
+        {
+            base.TestSearchAndValidateForTestUser(user);
+            base.TestAugmentationAgainst1RandomGroup(user);
+        }
+
+        [Test, TestCaseSource(typeof(TestEntitySourceManager), nameof(TestEntitySourceManager.GetSomeDisabledUsers), new object[] { 10 })]
+        public void TestDisabledUsers(TestUser user)
+        {
+            base.TestSearchAndValidateForTestUser(user);
+            base.TestAugmentationAgainst1RandomGroup(user);
+        }
+    }
+}