+++ Yvand.EntraCP/TEMPLATE/ADMIN/EntraCP/GlobalSettings.ascx.designer.cs
@@ -204,6 +204,15 @@ namespace Yvand.EntraClaimsProvider.Administration
         protected global::System.Web.UI.WebControls.DropDownList DdlUserGraphPropertyToDisplay;
 
         /// <summary>
+        /// ChkFilterAccountsEnabledOnly control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.CheckBox ChkFilterAccountsEnabledOnly;
+
+        /// <summary>
         /// DdlGroupClaimType control.
         /// </summary>
         /// <remarks>